!8 RELEASE版本1.1.9合并到master * chore 版本号更新为1.1.9-RELEASE * feat 增加管理员手动立即同步接口 * fix 修复了特定模式下，存在未同步文件信息时无法切换存储模式的问题 * docs 更新postman文档关于管理员添加用户的描述 * feat 特殊用户名不会允许被注册 * refactor 特殊用户名集中到User中编码 * refactor 移除SyncService无用导入，变量类型改用lombok自动推导 * chore 版本号升级为1.1.8-SNAPSHOT,数据同步性能优化 * refactor 移除ConfigureInitializer无用属性 * perf 优化同步速度，避免了重复的大量数据库查询 * feat NodeService获取NodeTree时会包含root节点 * feat NodeTree实现可迭代接口 * feat 增加节点树对象，一次性从数据库抓取整个目录树结构 * refactor 移除无用类 * refactor 同步服务拆分为数据模型，检测与数据处理模块 * refactor 对同步时的差异文件操作抽离出接口来执行 * remove 移除部分废弃的管理员接口 * docs 更新近期特性开发计划 * Merge branch 'develop' of https://gitee.com/xiaotao233/saltedfishcloud… * fix UNIQUE模式下文件被更改后的同步不会删除存储库的原文件和链接新文件 * feat 增加立即同步API * Merge branch 'master' into develop * chore 发布稳定版本1.1.7-RELEASE * chore 添加JAXB API依赖以支持Java9及以上版本运行支持 * fix 修复新用户注册后第一时间无法使用FTP私人网盘的问题 * fix 修复application.yml模板错误 * fix 修复初次运行程序时使用UNIQUE模式误判为旧版本升级的问题 * feat 使用非发行版时输出警告 * docs 更新README特性描述 * chore 构建和运行脚本自动识别打包目录下的jar文件 * style 移除无用import和修复Enum原始类型警告 * chore 系统版本从pom.xml中获取 * docs 添加FTP参数注意事项 * docs 存储切换说明改用start脚本用例 * chore 启动脚本支持附加命令行参数 * fix 修复版本升级时存储模式切换失败 * docs 添加README.md的命令行选项描述 * feat 添加命令行下的存储切换方式 * Merge branch 'filesystem-refactor' into develop * Merge branch 'filesystem-refactor' into develop * feat 添加版本升级控制 * Merge branch 'develop' into filesystem-refactor * !6 feat 在文件系统重构后实现FTP基本功能 * chore 版本号修改为1.1.0-SNAPSHOT * fix 修正bash启动脚本ftp参数错误 * feat 增加在Linux下路径的约束检查 * feat UNIQUE存储模式改用文件硬链接,本地将保持原文件目录结构

+++ src/main/java/com/xiaotao/saltedfishcloud/service/config/StoreTypeSwitch.java
@@ -66,6 +66,10 @@ public class StoreTypeSwitch {
                     if (file.isDir()) continue;
                     Path source = Paths.get(DiskConfig.uniquePathHandler.getStorePath(uid, p, file));
                     Path target = Paths.get(DiskConfig.rawPathHandler.getStorePath(uid, p, file));
+                    if (!Files.exists(source)) {
+                        log.warn("存储库文件丢失：" + file.getName() + " MD5:" + file.getMd5());
+                        continue;
+                    }
                     log.info("Copy file: " + source + " -> " + target);
                     Files.copy(source, target, StandardCopyOption.REPLACE_EXISTING);
                 }
@@ -92,6 +96,10 @@ public class StoreTypeSwitch {
                     if(fileInfo.isDir()) continue;
                     Path source = Paths.get(DiskConfig.rawPathHandler.getStorePath(user.getId(), path, fileInfo));
                     String target = DiskConfig.uniquePathHandler.getStorePath(user.getId(), path, fileInfo);
+                    if (!Files.exists(source)) {
+                        log.warn("未同步的文件：" + path + "/" + fileInfo.getName());
+                        continue;
+                    }
                     log.info("Copy file: " + source + " -> " + target);
                     storeService.store(user.getId(), Files.newInputStream(source), path, fileInfo);
                 }
